refactor: isolate configuration code
