+++ .gitignore
@@ -9,6 +9,7 @@ logs
 .tmp-earthly-out
 .vscode/settings.json
 .ruff_cache
+data/
 
 ################################################################################
 # Rust. Generated by Cargo                                                     #
